+++ plutarch-testlib/Plutarch/Test/Golden.hs
@@ -20,9 +20,15 @@ import Data.Int (Int64)
 import Data.Text (Text)
 import Data.Text qualified as Text
 import Data.Text.Encoding qualified as Encoding
-import Plutarch (ClosedTerm, Config (Tracing), LogLevel (LogInfo), Script, TracingMode (DetTracing), compile)
 import Plutarch.Evaluate (EvalError, evalScript)
 import Plutarch.Internal.Other (printScript)
+import Plutarch.Internal.Term (
+  ClosedTerm,
+  Config (Tracing),
+  LogLevel (LogInfo),
+  TracingMode (DetTracing),
+  compile,
+ )
 import Plutarch.Script (Script (unScript))
 import PlutusLedgerApi.Common (serialiseUPLC)
 import PlutusLedgerApi.V1 (ExBudget (ExBudget), ExCPU, ExMemory)