+++ plutarch-testlib/Plutarch/Test/Unit.hs
@@ -11,14 +11,14 @@ module Plutarch.Test.Unit (
 
 import Data.Text (Text)
 import Data.Text qualified as Text
-import Plutarch (
+import Plutarch.Evaluate (EvalError, evalScript)
+import Plutarch.Internal.Other (printScript)
+import Plutarch.Internal.Term (
   Config (NoTracing, Tracing),
   LogLevel (LogDebug),
   TracingMode (DetTracing),
   compile,
-  printScript,
  )
-import Plutarch.Evaluate (EvalError, evalScript)
 import Plutarch.Prelude
 import Test.Tasty (TestName, TestTree)
 import Test.Tasty.HUnit (assertEqual, assertFailure, testCase)