+++ plutarch-testlib/test/Plutarch/Test/Suite/Plutarch/Scripts.hs
@@ -8,9 +8,14 @@ import Data.Coerce (coerce)
 import Data.Text (Text)
 import Data.Text qualified as Text
 import Data.Text.Encoding qualified as Encoding
-import Plutarch (Config (Tracing), LogLevel (LogInfo), TracingMode (DetTracing), compile)
 import Plutarch.Builtin (pasByteStr)
 import Plutarch.Crypto (pverifyEd25519Signature)
+import Plutarch.Internal.Term (
+  Config (Tracing),
+  LogLevel (LogInfo),
+  TracingMode (DetTracing),
+  compile,
+ )
 import Plutarch.LedgerApi.V3 (
   PPubKeyHash,
   PScriptContext,